requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
absl-py==0.11.0
astroid==2.4.2
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
demjson==2.2.4
Flask==1.1.2
fsspec==0.8.4
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
gunicorn==20.0.4
healthcheck==1.3.3
idna==2.10
imutils==0.5.3
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jtutils==0.0.8
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
nose==1.3.7
numpy==1.19.2
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
Pillow==8.0.0
protobuf==3.14.0
py-healthcheck==1.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pytorch-lightning==1.0.8
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
toml==0.10.1
torch==1.6.0+cpu
torchvision==0.7.0+cpu
tqdm==4.54.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
